requirements.txt
openai==1.3.0
sqlite3
sqlalchemy==2.0.23
langchain==0.1.0
langchain-openai==0.0.2
python-dotenv==1.0.0